+++ lib/playercenter-backend/utils.rb
@@ -9,4 +9,3 @@ module Playercenter::Backend
     end
   end
 end
-